+++ cranko-release-info-v1
[[projects]]
qnames = ["tectonic_xdv", "cargo"]
version = "0.2.2"
age = 0

[[projects]]
qnames = ["tectonic_errors", "cargo"]
version = "0.2.1"
age = 0

[[projects]]
qnames = ["tectonic_xetex_format", "cargo"]
version = "0.3.1"
age = 0

[[projects]]
qnames = ["tectonic_status_base", "cargo"]
version = "0.2.1"
age = 0

[[projects]]
qnames = ["tectonic_io_base", "cargo"]
version = "0.4.2"
age = 0

[[projects]]
qnames = ["tectonic_geturl", "cargo"]
version = "0.3.2"
age = 0

[[projects]]
qnames = ["tectonic_docmodel", "cargo"]
version = "0.2.1"
age = 0

[[projects]]
qnames = ["tectonic_dep_support", "cargo"]
version = "0.1.1"
age = 0

[[projects]]
qnames = ["tectonic_cfg_support", "cargo"]
version = "0.1.4"
age = 11

[[projects]]
qnames = ["tectonic_bundles", "cargo"]
version = "0.3.1"
age = 0

[[projects]]
qnames = ["tectonic_bridge_icu", "cargo"]
version = "0.2.0"
age = 32

[[projects]]
qnames = ["tectonic_bridge_graphite2", "cargo"]
version = "0.2.2"
age = 0

[[projects]]
qnames = ["tectonic_bridge_harfbuzz", "cargo"]
version = "0.2.7"
age = 2

[[projects]]
qnames = ["tectonic_bridge_freetype2", "cargo"]
version = "0.2.0"
age = 32

[[projects]]
qnames = ["tectonic_bridge_flate", "cargo"]
version = "0.1.7"
age = 0

[[projects]]
qnames = ["tectonic_bridge_core", "cargo"]
version = "0.3.2"
age = 0

[[projects]]
qnames = ["tectonic_xetex_layout", "cargo"]
version = "0.2.2"
age = 0

[[projects]]
qnames = ["tectonic_pdf_io", "cargo"]
version = "0.4.1"
age = 0

[[projects]]
qnames = ["tectonic_engine_xetex", "cargo"]
version = "0.4.2"
age = 0

[[projects]]
qnames = ["tectonic_engine_xdvipdfmx", "cargo"]
version = "0.4.1"
age = 0

[[projects]]
qnames = ["tectonic_engine_spx2html", "cargo"]
version = "0.3.0"
age = 0

[[projects]]
qnames = ["tectonic_engine_bibtex", "cargo"]
version = "0.1.5"
age = 0

[[projects]]
qnames = ["tectonic", "cargo"]
version = "0.12.0"
age = 1

+++

+++ crates/xetex_format/src/cshash.rs
@@ -190,7 +190,7 @@ impl ControlSeqHash {
                 // The 1 here is formally ACTIVE_BASE
                 return Some(format!(
                     "[active character {}]",
-                    crate::format::fmt_usv(p as i32 - 1)
+                    crate::format::fmt_usv(p - 1)
                 ));
             }
         }
